+++ src/blocks.js
@@ -11403,10 +11403,6 @@ MultiArgMorph.prototype = new ArgMorph();
 MultiArgMorph.prototype.constructor = MultiArgMorph;
 MultiArgMorph.uber = ArgMorph.prototype;
 
-// MultiArgMorph preferences settings:
-
-// MultiArgMorph.prototype.isCachingInputs = true; // commented out for now
-
 // MultiArgMorph instance creation:
 
 function MultiArgMorph(
@@ -11525,7 +11521,6 @@ MultiArgMorph.prototype.setContents = function (anArray) {
             inputs[i].setContents(anArray[i]);
         }
     }
-    this.cachedInputs = null;
 };
 
 // MultiArgMorph hiding and showing:
@@ -11686,7 +11681,6 @@ MultiArgMorph.prototype.addInput = function (contents) {
     this.children.splice(idx, 0, newPart);
     newPart.fixLayout();
     this.fixLayout();
-    this.cachedInputs = null;
     return newPart;
 };
 
@@ -11704,7 +11698,6 @@ MultiArgMorph.prototype.removeInput = function () {
         }
     }
     this.fixLayout();
-    this.cachedInputs = null;
 };
 
 MultiArgMorph.prototype.is3ArgRingInHOF = function () {